+++ Convexus-Tests/src/main/java/exchange/convexus/testtokens/Usdc.java
@@ -56,8 +56,8 @@ public class Usdc {
     }
 
     @External(readonly=true)
-    public int decimals() {
-        return decimals;
+    public BigInteger decimals() {
+        return BigInteger.valueOf(decimals);
     }
 
     @External(readonly=true)
