+++ javaee-tokens/src/test/java/com/iconloop/score/token/irc2/IRC2BasicTest.java
@@ -66,7 +66,7 @@ class IRC2BasicTest extends TestBase {
 
     @Test
     void decimals() {
-        assertEquals(decimals, tokenScore.call("decimals"));
+        assertEquals(decimals, ((BigInteger) tokenScore.call("decimals")).intValue());
     }
 
     @Test
